+++ src/init.ts
@@ -3,8 +3,6 @@
 import { commands } from 'vscode';
 import Git from './Git';
 
-(async function(): Promise<void> {
-    if (await (new Git()).isGitRepository()) {
-        commands.executeCommand('setContext', 'isGitRepository', true);
-    }
-})();
+(new Git()).isGitRepository().then(() => {
+    commands.executeCommand('setContext', 'isGitRepository', true);
+});