+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [[0.6.0] 2017-10-01](https://github.com/arturock/vscode-gitstash/tree/v0.6.0)
+- Explorer Tree
+  - Add buttons
+- Settings
+  - Autoclear log
+  - show / hide the explorer buttons
+- Commands
+  - Reload tree
+  - show / hide explorer
+Commands are now shown only when on a git repository
+
 ## [[0.5.0] 2017-10-01](https://github.com/arturock/vscode-gitstash/tree/v0.5.0)
  - Add missing commands
    - Save simple, keep index, include untracked, stash all